revise `quote` and `quasiquote`
Allow `quote` to be used in ways that generate typable lists,
and make the set of S-expression values (produced by `quasiquote`)
distinct from other values so that an S-expression is always
printed with a backquote.

This changes make value printing far more consistent the expression
grammar. A drawback is that the distinct representation of
S-expressions (a wrapper around a raw S-expression) is exposed to
non-`plait` programs.