support type declarations among definitions; adjust REPL typechecking
Support type declarations separate from function implementations,
similar to Typed Racket and Haskell:

 (inc : (Number -> Number))
 (define (inc x)
   (+ x 1))

Adding support fo type declarations in the REPL exposed that a
redefinition in the REPL could change an identifier's type. This
commit repairs that problem by disallowing a type change in a
redefinition. A consequence of this change is that an identifier with
a polymorphic types cannot be redefined, because the type checker
cannot unify polymorphic types.